mqtt-bridge/app/requirements.txt
influxdb
typing
paho-mqtt